Refactor sugerencias de compra para usar Integer en lugar de Long para el ID de sede

+++ backend-inventario/src/main/java/com/inventario/backend_inventario/Controller/ProductoController.java
@@ -70,7 +70,7 @@ public class ProductoController {
     }
 
     @GetMapping("/sugerencias/reabastecer")
-    public ResponseEntity<List<SugerenciaCompraDto>> sugerirCompra(@RequestParam Long idSede) {
+    public ResponseEntity<List<SugerenciaCompraDto>> sugerirCompra(@RequestParam Integer idSede) {
         return ResponseEntity.ok(productoService.obtenerSugerenciasReabastecimiento(idSede));
     }
 }
